Refactored structure of controllers
