+++ geoportaal-beheer/app/controllers/MetadataDC.java
@@ -9,8 +9,10 @@ import static models.QMdSubject.mdSubject;
 import static models.QMetadata.metadata;
 import static models.QRights.rights;
 import static models.QRightsLabel.rightsLabel;
+import static models.QStatus.status;
 import static models.QSubject.subject;
 import static models.QSubjectLabel.subjectLabel;
+import static models.QSupplier.supplier;
 import static models.QTypeInformation.typeInformation;
 import static models.QTypeInformationLabel.typeInformationLabel;
 import static models.QUseLimitation.useLimitation;
@@ -36,7 +38,7 @@ import play.mvc.Controller;
 import play.mvc.Http.MultipartFormData.FilePart;
 import play.mvc.Result;
 
-public class Add extends Controller {
+public class MetadataDC extends Controller {
 	@Inject Database db;
 	
 	public Result renderCreateForm() {
@@ -203,6 +205,77 @@ public class Add extends Controller {
 		}
 		
 		return redirect(controllers.routes.Index.index());
+	}
+	
+	public Result renderEditForm(String datasetUuid) {
+		Boolean create = false;
+		
+		Integer datasetId = db.queryFactory
+				.from(metadata)
+				.select(metadata.id)
+				.where(metadata.uuid.eq(datasetUuid))
+				.fetchFirst();
+		
+		Tuple datasetRow = db.queryFactory.select(metadata.id, metadata.uuid, metadata.location, metadata.fileId, metadata.title, 
+				metadata.description, metadata.typeInformation, metadata.creator, metadata.creatorOther, metadata.rights, metadata.useLimitation,
+				metadata.mdFormat, metadata.source, metadata.dateSourceCreation, metadata.dateSourcePublication, metadata.dateSourceRevision,
+				metadata.dateSourceValidFrom, metadata.dateSourceValidUntil, creator.name)
+    			.from(metadata)
+    			.join(creator).on(metadata.creator.eq(creator.id))
+    			.where(metadata.id.eq(datasetId))
+    			.fetchFirst();
+    	
+    	List<Tuple> subjectsDataset = db.queryFactory.select(mdSubject.all())
+    			.from(mdSubject)
+    			.where(mdSubject.metadataId.eq(datasetId))
+    			.fetch();
+    	
+    	List<Tuple> attachmentsDataset = db.queryFactory.select(mdAttachment.all())
+    			.from(mdAttachment)
+    			.where(mdAttachment.metadataId.eq(datasetId))
+    			.fetch();
+    	
+    	List<Tuple> typeInformationList = db.queryFactory
+    			.select(typeInformation.id, typeInformation.name, typeInformationLabel.label)
+        		.from(typeInformation)
+        		.join(typeInformationLabel).on(typeInformation.id.eq(typeInformationLabel.typeInformationId))
+        		.fetch();
+        	
+    	List<Tuple> creatorsList = db.queryFactory
+	    		.select(creator.id, creator.name, creatorLabel.label)
+	        	.from(creator)
+	        	.join(creatorLabel).on(creator.id.eq(creatorLabel.creatorId))
+	        	.fetch();
+    	
+    	List<Tuple> rightsList = db.queryFactory
+    			.select(rights.id, rights.name, rightsLabel.label)
+            	.from(rights)
+            	.join(rightsLabel).on(rights.id.eq(rightsLabel.rightsId))
+            	.fetch();
+    	
+    	List<Tuple> useLimitationList = db.queryFactory
+    			.select(useLimitation.id, useLimitation.name, useLimitationLabel.label)
+            	.from(useLimitation)
+            	.join(useLimitationLabel).on(useLimitation.id.eq(useLimitationLabel.useLimitationId))
+            	.fetch();
+    	
+    	List<Tuple> mdFormatList = db.queryFactory
+    			.select(mdFormat.id, mdFormat.name, mdFormatLabel.label)
+            	.from(mdFormat)
+            	.join(mdFormatLabel).on(mdFormat.id.eq(mdFormatLabel.mdFormatId))
+            	.fetch();
+    	
+    	List<Tuple> subjectList = db.queryFactory
+    			.select(subject.id, subject.name, subjectLabel.label)
+            	.from(subject)
+            	.join(subjectLabel).on(subject.id.eq(subjectLabel.subjectId))
+            	.fetch();
+        
+        SimpleDateFormat sdfUS = new SimpleDateFormat("yyyy-MM-dd");
+        SimpleDateFormat sdfLocal = new SimpleDateFormat("dd-MM-yyyy");
+        
+        return ok(views.html.form.render(create, "", "", datasetRow, subjectsDataset, attachmentsDataset, typeInformationList, creatorsList, 
+    			rightsList, useLimitationList, mdFormatList, sdfUS, sdfLocal, subjectList));
 	}
 	
 	public Timestamp nullCheckDate(Date date) {
