+++ geoportaal-beheer/app/views/index.scala.html
@@ -125,7 +125,7 @@
 							<a>
 								<span class="action-button glyphicon glyphicon-eye-open" title="Metadata bekijken"></span>
 							</a>
-							<a href="@routes.Edit.renderEditForm(datasetRow.get(metadata.uuid))">
+							<a href="@routes.MetadataDC.renderEditForm(datasetRow.get(metadata.uuid))">
 								<span class="action-button glyphicon glyphicon-edit" title="Bestand wijzigen"></span>
 							</a>
 						</td>
