+++ geoportaal-beheer/app/views/main.scala.html
@@ -26,7 +26,7 @@
 				<div class="toptaak-element">
 					<div class="toptaak-outer">
 						<div class="@if(title == "Toevoegen") {toptaak-inner-active} else {toptaak-inner}">
-							<a class="tabs" href="@routes.Add.renderCreateForm()">
+							<a class="tabs" href="@routes.MetadataDC.renderCreateForm()">
 								<span class="toptaak-titel">
 									Toevoegen
 								</span>
