+++ geoportaal-beheer/conf/routes
@@ -7,13 +7,13 @@ GET     /                           controllers.Index.index()
 GET		/user/login					controllers.User.login(r: String)
 GET		/user/logout				controllers.User.logout()
 POST	/user/authenticate			controllers.User.authenticate()
-GET		/add						controllers.Add.renderCreateForm()
-POST	/submit						controllers.Add.createSubmit()
-GET		/edit/:datasetUuid			controllers.Edit.renderEditForm(datasetUuid: String)
-GET		/editStatus/:datasetId		controllers.Edit.changeStatus(datasetId: Integer, status: String)
-GET		/editSupplier/:datasetId	controllers.Edit.changeSupplier(datasetId: Integer, supplier: String)
-GET		/cancel						controllers.Add.cancel()
-GET		/delete/:datasetId			controllers.Delete.delete(datasetId: Integer)
+GET		/add						controllers.MetadataDC.renderCreateForm()
+POST	/submit						controllers.MetadataDC.createSubmit()
+GET		/edit/:datasetUuid			controllers.MetadataDC.renderEditForm(datasetUuid: String)
+GET		/editStatus/:datasetId		controllers.Index.changeStatus(datasetId: Integer, status: String)
+GET		/editSupplier/:datasetId	controllers.Index.changeSupplier(datasetId: Integer, supplier: String)
+GET		/cancel						controllers.MetadataDC.cancel()
+GET		/delete/:datasetId			controllers.Index.deleteMetadata(datasetId: Integer)
 GET		/routes.js					controllers.Index.jsRoutes()
 
 # Map static resources from the /public folder to the /assets URL path
